+++ src/log.js
@@ -5,8 +5,10 @@
  *
  * To use the logger, simply call HHM.log.{info,warn,error,…}().
  *
- * After the room was started, messages can be logged to the room using
- * `HHM.log.toRoom(message, level)`.
+ * After the room was started, messages can be logged using
+ * `room.log(message, level)`.
+ *
+ * Levels can be accessed via HHM.log.level.*
  *
  * @see https://github.com/pimterry/loglevel
  */
@@ -18,6 +20,16 @@ module.exports = function Constructor() {
   loglevelMessagePrefix(log, {
     prefixes: [`level`],
     staticPrefixes: [`HHM`],
+  });
+
+  $.extend(log, {
+    level: {
+      TRACE: `trace`,
+      DEBUG: `debug`,
+      INFO: `info`,
+      WARN: `warn`,
+      ERROR: `error`
+    }
   });